+++ src/EntityGraphQL/Schema/ISchemaProvider.cs
@@ -29,7 +29,7 @@ namespace EntityGraphQL.Schema
         ISchemaType AddUnion(Type type, string name, string? description);
         SchemaType<TBaseType> AddInputType<TBaseType>(string name, string? description);
         ISchemaType AddInputType(Type type, string name, string? description);
-        void AddMutationsFrom<TType>(SchemaBuilderMethodOptions? options = null) where TType : class;
+        void AddMutationsFrom<TType>(SchemaBuilderOptions? options = null) where TType : class;
         ISchemaType AddScalarType(Type clrType, string gqlTypeName, string? description);
         SchemaType<TType> AddScalarType<TType>(string gqlTypeName, string? description);
         SchemaType<TBaseType> AddType<TBaseType>(string name, string? description);